[color-contrast-tool] combine files
