+++ __docs__/wonder-blocks-tokens/core-color-contrast.stories.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import * as React from "react";
 import {StyleSheet} from "aphrodite";
 import {Meta, StoryObj} from "@storybook/react-vite";
@@ -7,35 +8,30 @@ import xCircleFillIcon from "@phosphor-icons/core/fill/x-circle-fill.svg";
 
 import {View} from "@khanacademy/wonder-blocks-core";
 import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
+import {ThemeSwitcher} from "@khanacademy/wonder-blocks-theming";
 import {BodyText} from "@khanacademy/wonder-blocks-typography";
 import {
     border,
+    font,
     mix,
     semanticColor,
     sizing,
 } from "@khanacademy/wonder-blocks-tokens";
 
+import {color as thunderblocksPrimitives} from "../../packages/wonder-blocks-tokens/src/theme/semantic/internal/primitive-color-thunderblocks";
 import {
     flattenNestedTokens,
     maybeGetCssVariableInfo,
 } from "../components/tokens-util";
 
 /**
- * A reference grid that pairs every `semanticColor.core.background` token with
- * every `semanticColor.core.foreground` token and reports whether the pair
- * meets WCAG **Level AA** contrast.
+ * Stories that visualize WCAG **Level AA** contrast between every
+ * `semanticColor.core.background` token and every
+ * `semanticColor.core.foreground` token.
  *
- * Background tokens are listed down the left side; foreground tokens run along
- * the bottom. Tokens with alpha (e.g. `core.background.neutral.subtle`) are
- * composited before the contrast is calculated:
- *
- * - The background is composited over `core.background.base.default` (the
- *   surface most cards/pages sit on) so the visible color includes the page
- *   underneath.
- * - The foreground is then composited over the resolved background, since
- *   that's what a reader's eye actually sees.
- *
- * Cells pass at the **AA normal text** threshold (≥ 4.5 : 1).
+ * Tokens with alpha (e.g. `core.background.neutral.subtle`) are composited
+ * over `core.background.base.default` (per theme) before contrast is
+ * computed, so the visible color drives the decision.
  */
 export default {
     title: "Internal / Semantic Color Contrast",
@@ -44,6 +40,12 @@ export default {
     },
     tags: ["!manifest"],
 } as Meta;
+
+// ---------------------------------------------------------------------------
+// Shared utilities
+// ---------------------------------------------------------------------------
+
+const AA_THRESHOLD = 4.5;
 
 type FlatToken = {label: string; cssVar: string};
 
@@ -164,13 +166,41 @@ function formatRatio(ratio: number): string {
     return ratio.toFixed(2).replace(/\.?0+$/, "");
 }
 
-type Resolved = {
+function capitalize(value: string): string {
+    return value.charAt(0).toUpperCase() + value.slice(1);
+}
+
+function normalizeColorString(value: string): string {
+    return (value || "").trim().toLowerCase().replace(/\s+/g, "");
+}
+
+/**
+ * Reverse map from a primitive's color value to its primitive name (e.g.
+ * "#ffffff" → "white_100"). Both Thunderblocks and Syl-Dark draw from this
+ * same primitive palette.
+ */
+const PRIMITIVE_BY_VALUE = new Map<string, string>(
+    Object.entries(thunderblocksPrimitives).map(([name, value]) => [
+        normalizeColorString(value),
+        name,
+    ]),
+);
+
+function lookupPrimitiveName(value: string | undefined): string | null {
+    if (!value) {
+        return null;
+    }
+    return PRIMITIVE_BY_VALUE.get(normalizeColorString(value)) ?? null;
+}
+
+type ResolvedTheme = {
     backgrounds: Record<string, string>;
     foregrounds: Record<string, string>;
     pageBase: string;
 };
 
-function resolveAll(): Resolved {
+/** Resolve every bg / fg token against the active theme on the document. */
+function resolveActiveTheme(): ResolvedTheme {
     const pageBase =
         maybeGetCssVariableInfo(semanticColor.core.background.base.default)
             .value || "#ffffff";
@@ -191,11 +221,37 @@ function resolveAll(): Resolved {
     };
 }
 
-function useResolvedTokens(): Resolved | null {
-    const [resolved, setResolved] = React.useState<Resolved | null>(null);
+/** Resolve every bg / fg token using the CSS-var scope of `el`. */
+function resolveScopedTheme(el: HTMLElement | null): ResolvedTheme | null {
+    if (!el) {
+        return null;
+    }
+    const cs = getComputedStyle(el);
+    const resolve = (cssVar: string) => {
+        const match = cssVar.match(/^var\((--[^,)]+)/);
+        if (!match) {
+            return cssVar;
+        }
+        return cs.getPropertyValue(match[1]).trim() || cssVar;
+    };
+    return {
+        backgrounds: Object.fromEntries(
+            backgroundTokens.map(({label, cssVar}) => [label, resolve(cssVar)]),
+        ),
+        foregrounds: Object.fromEntries(
+            foregroundTokens.map(({label, cssVar}) => [label, resolve(cssVar)]),
+        ),
+        pageBase:
+            resolve(semanticColor.core.background.base.default) || "#ffffff",
+    };
+}
+
+/** Read the current document-level theme and re-resolve when it changes. */
+function useResolvedActiveTheme(): ResolvedTheme | null {
+    const [resolved, setResolved] = React.useState<ResolvedTheme | null>(null);
 
     React.useEffect(() => {
-        const update = () => setResolved(resolveAll());
+        const update = () => setResolved(resolveActiveTheme());
         update();
 
         const observer = new MutationObserver(update);
@@ -215,11 +271,52 @@ function useResolvedTokens(): Resolved | null {
     return resolved;
 }
 
-const AA_THRESHOLD = 4.5;
+type ResolvedPair = {
+    bgLabel: string;
+    fgLabel: string;
+    bgCssVar: string;
+    fgCssVar: string;
+    bgRaw: string;
+    fgRaw: string;
+    bgVisible: string;
+    fgVisible: string;
+    ratio: number;
+};
 
-function capitalize(value: string): string {
-    return value.charAt(0).toUpperCase() + value.slice(1);
+function buildResolvedPairs(theme: ResolvedTheme): Array<ResolvedPair> {
+    const pairs: Array<ResolvedPair> = [];
+    for (const bg of backgroundTokens) {
+        const bgRaw = theme.backgrounds[bg.label];
+        if (isTransparentValue(bgRaw)) {
+            continue;
+        }
+        const bgVisible = compositeOver(bgRaw, theme.pageBase);
+        for (const fg of foregroundTokens) {
+            const fgRaw = theme.foregrounds[fg.label];
+            const fgVisible = compositeOver(fgRaw, bgVisible);
+            const ratio = contrastRatio(bgVisible, fgVisible);
+            if (ratio == null) {
+                continue;
+            }
+            pairs.push({
+                bgLabel: bg.label,
+                fgLabel: fg.label,
+                bgCssVar: bg.cssVar,
+                fgCssVar: fg.cssVar,
+                bgRaw,
+                fgRaw,
+                bgVisible,
+                fgVisible,
+                ratio,
+            });
+        }
+    }
+    return pairs;
 }
+
+// ---------------------------------------------------------------------------
+// Story 1: Foreground × Background grid
+// ---------------------------------------------------------------------------
 
 function ContrastCell({
     bgRaw,
@@ -322,9 +419,11 @@ const FIRST_CELL_COL = 3;
 const ROW_TOP_HEADING = 1;
 const ROW_FIRST_BG = 2;
 
+const CELL_SIZE = 64;
+
 export const ForegroundOnBackground: StoryObj = {
     render: function Render() {
-        const resolved = useResolvedTokens();
+        const resolved = useResolvedActiveTheme();
 
         if (!resolved) {
             return (
@@ -344,7 +443,6 @@ export const ForegroundOnBackground: StoryObj = {
 
         const elements: React.ReactElement[] = [];
 
-        // Top-left "Core Background" heading
         elements.push(
             <View
                 key="bg-heading"
@@ -367,11 +465,9 @@ export const ForegroundOnBackground: StoryObj = {
             </View>,
         );
 
-        // BG rows
         let bgRow = ROW_FIRST_BG;
         for (const group of backgroundGroups) {
             const groupSize = group.variants.length;
-            // Outer category label (rowSpan)
             elements.push(
                 <View
                     key={`bg-cat-${group.name}`}
@@ -402,7 +498,6 @@ export const ForegroundOnBackground: StoryObj = {
                 const resolvedBg = backgrounds[fullLabel];
                 const transparent = isTransparentValue(resolvedBg);
 
-                // Variant swatch
                 elements.push(
                     <View
                         key={`bg-swatch-${fullLabel}`}
@@ -419,7 +514,6 @@ export const ForegroundOnBackground: StoryObj = {
                     </View>,
                 );
 
-                // Cells
                 if (!transparent) {
                     foregroundTokens.forEach((fg, j) => {
                         const col = FIRST_CELL_COL + j;
@@ -463,7 +557,6 @@ export const ForegroundOnBackground: StoryObj = {
             );
         });
 
-        // FG outer category labels (with col span)
         let fgCol = FIRST_CELL_COL;
         for (const group of foregroundGroups) {
             const groupSize = group.variants.length;
@@ -491,7 +584,6 @@ export const ForegroundOnBackground: StoryObj = {
             fgCol += groupSize;
         }
 
-        // "Foreground" heading centered under the FG axis
         elements.push(
             <View
                 key="fg-heading"
@@ -547,11 +639,376 @@ export const SylDarkForegroundOnBackground = {
     },
 };
 
-const CELL_SIZE = 64;
+// ---------------------------------------------------------------------------
+// Story 2: Cross-theme passing pairs (Thunderblocks vs Syl-Dark)
+// ---------------------------------------------------------------------------
+
+function TokenRow({
+    kind,
+    tokenName,
+    value,
+}: {
+    kind: "bg" | "fg";
+    tokenName: string;
+    value: string | undefined;
+}) {
+    const trimmed = value?.trim() ?? "";
+    const primitive = lookupPrimitiveName(trimmed);
+    return (
+        <>
+            <BodyText
+                tag="span"
+                size="xsmall"
+                weight="bold"
+                style={styles.tokenCellKind}
+            >
+                {kind}
+            </BodyText>
+            <BodyText tag="span" size="xsmall" style={styles.tokenCellName}>
+                {tokenName}
+            </BodyText>
+            <BodyText tag="span" size="xsmall" style={styles.tokenCellMono}>
+                {primitive ?? "—"}
+            </BodyText>
+            <BodyText tag="span" size="xsmall" style={styles.tokenCellMono}>
+                {trimmed || "—"}
+            </BodyText>
+        </>
+    );
+}
+
+function ThemePreview({
+    bgCssVar,
+    fgCssVar,
+    bgLabel,
+    fgLabel,
+    bgValue,
+    fgValue,
+    ratio,
+    highlightFailing = false,
+}: {
+    bgCssVar: string;
+    fgCssVar: string;
+    bgLabel: string;
+    fgLabel: string;
+    /** The actual color string the bg CSS var resolves to in this theme. */
+    bgValue: string | undefined;
+    /** The actual color string the fg CSS var resolves to in this theme. */
+    fgValue: string | undefined;
+    ratio: number | null;
+    /** When true, draws the bright-red outline if the pair fails AA. */
+    highlightFailing?: boolean;
+}) {
+    const passes = ratio != null && ratio >= AA_THRESHOLD;
+    return (
+        <View
+            style={[
+                styles.previewTile,
+                highlightFailing && !passes && styles.previewTileFailing,
+            ]}
+        >
+            <View
+                style={[
+                    styles.previewSwatch,
+                    {backgroundColor: bgCssVar, color: fgCssVar},
+                ]}
+            >
+                <BodyText
+                    tag="span"
+                    size="medium"
+                    weight="bold"
+                    style={styles.previewSampleHeading}
+                >
+                    Aa Bb Cc
+                </BodyText>
+                <BodyText
+                    tag="span"
+                    size="small"
+                    style={styles.previewSampleBody}
+                >
+                    The quick brown fox jumps.
+                </BodyText>
+            </View>
+            <View style={styles.previewMeta}>
+                <View style={styles.previewMetaRow}>
+                    <PhosphorIcon
+                        icon={passes ? checkCircleFillIcon : xCircleFillIcon}
+                        size="small"
+                        color={
+                            passes
+                                ? semanticColor.core.foreground.success.default
+                                : semanticColor.core.foreground.neutral.subtle
+                        }
+                        aria-hidden={true}
+                    />
+                    <BodyText tag="span" size="xsmall">
+                        {ratio == null
+                            ? "no value"
+                            : `${formatRatio(ratio)} : 1`}{" "}
+                        · {passes ? "passes AA" : "fails AA"}
+                    </BodyText>
+                </View>
+                <View style={styles.tokenTable}>
+                    <TokenRow
+                        kind="bg"
+                        tokenName={`core.background.${bgLabel}`}
+                        value={bgValue}
+                    />
+                    <TokenRow
+                        kind="fg"
+                        tokenName={`core.foreground.${fgLabel}`}
+                        value={fgValue}
+                    />
+                </View>
+            </View>
+        </View>
+    );
+}
+
+export const PassingPairsThunderblocks: StoryObj = {
+    name: "Thunderblocks ↔ SylDark",
+    parameters: {
+        chromatic: {disableSnapshot: true},
+    },
+    render: function Render() {
+        const tbRef = React.useRef<HTMLDivElement>(null);
+        const sdRef = React.useRef<HTMLDivElement>(null);
+        const [themes, setThemes] = React.useState<{
+            tb: ResolvedTheme;
+            sd: ResolvedTheme;
+        } | null>(null);
+
+        React.useEffect(() => {
+            const tb = resolveScopedTheme(tbRef.current);
+            const sd = resolveScopedTheme(sdRef.current);
+            if (tb && sd) {
+                setThemes({tb, sd});
+            }
+        }, []);
+
+        const passing = React.useMemo(() => {
+            if (!themes) {
+                return [];
+            }
+            return buildResolvedPairs(themes.tb).filter(
+                (pair) => pair.ratio >= AA_THRESHOLD,
+            );
+        }, [themes]);
+
+        const sdRatios = React.useMemo(() => {
+            if (!themes) {
+                return new Map<string, number | null>();
+            }
+            const map = new Map<string, number | null>();
+            for (const pair of passing) {
+                const bgRaw = themes.sd.backgrounds[pair.bgLabel];
+                const fgRaw = themes.sd.foregrounds[pair.fgLabel];
+                if (isTransparentValue(bgRaw)) {
+                    map.set(`${pair.bgLabel}__${pair.fgLabel}`, null);
+                    continue;
+                }
+                const bgVisible = compositeOver(bgRaw, themes.sd.pageBase);
+                const fgVisible = compositeOver(fgRaw, bgVisible);
+                map.set(
+                    `${pair.bgLabel}__${pair.fgLabel}`,
+                    contrastRatio(bgVisible, fgVisible),
+                );
+            }
+            return map;
+        }, [themes, passing]);
+
+        const sdFailingCount = React.useMemo(() => {
+            let n = 0;
+            for (const ratio of sdRatios.values()) {
+                if (ratio == null || ratio < AA_THRESHOLD) {
+                    n++;
+                }
+            }
+            return n;
+        }, [sdRatios]);
+
+        return (
+            <View style={styles.crossThemeWrapper}>
+                {/* Hidden helpers used to read each theme's CSS variables. */}
+                <View style={styles.hiddenProbe}>
+                    <ThemeSwitcher theme="thunderblocks">
+                        <div ref={tbRef} />
+                    </ThemeSwitcher>
+                    <ThemeSwitcher theme="syl-dark">
+                        <div ref={sdRef} />
+                    </ThemeSwitcher>
+                </View>
+
+                <View style={styles.summary}>
+                    <BodyText size="medium" weight="bold">
+                        Passing AA pairs in thunderblocks
+                    </BodyText>
+                    {themes ? (
+                        <>
+                            <BodyText size="small">
+                                {passing.length} of{" "}
+                                {backgroundTokens.length *
+                                    foregroundTokens.length}{" "}
+                                background × foreground combinations meet ≥ 4.5
+                                : 1 in thunderblocks. Each row shows the same
+                                pair in syl-dark for comparison.
+                            </BodyText>
+                            <BodyText size="small">
+                                {sdFailingCount} of those {passing.length} pair
+                                {passing.length === 1 ? "" : "s"} fall below 4.5
+                                : 1 (or use a transparent background) in
+                                syl-dark.
+                            </BodyText>
+                        </>
+                    ) : (
+                        <BodyText size="small">
+                            Resolving theme tokens…
+                        </BodyText>
+                    )}
+                </View>
+
+                <View style={styles.split}>
+                    <ThemeSwitcher theme="thunderblocks">
+                        <View style={styles.lightPanel}>
+                            <View style={styles.headerCell}>
+                                <BodyText
+                                    tag="span"
+                                    size="medium"
+                                    weight="bold"
+                                    style={styles.headerText}
+                                >
+                                    Thunderblocks (SYL Light)
+                                </BodyText>
+                            </View>
+                            {passing.map((pair) => (
+                                <ThemePreview
+                                    key={`${pair.bgLabel}__${pair.fgLabel}`}
+                                    bgCssVar={pair.bgCssVar}
+                                    fgCssVar={pair.fgCssVar}
+                                    bgLabel={pair.bgLabel}
+                                    fgLabel={pair.fgLabel}
+                                    bgValue={pair.bgRaw}
+                                    fgValue={pair.fgRaw}
+                                    ratio={pair.ratio}
+                                />
+                            ))}
+                        </View>
+                    </ThemeSwitcher>
+                    <ThemeSwitcher theme="syl-dark">
+                        <View style={styles.darkPanel}>
+                            <View style={styles.darkRow}>
+                                <View style={styles.headerCell}>
+                                    <BodyText
+                                        tag="span"
+                                        size="medium"
+                                        weight="bold"
+                                        style={styles.headerText}
+                                    >
+                                        SYL Dark
+                                    </BodyText>
+                                </View>
+                                <View
+                                    style={[
+                                        styles.headerCell,
+                                        styles.failingSlot,
+                                    ]}
+                                >
+                                    <BodyText
+                                        tag="span"
+                                        size="medium"
+                                        weight="bold"
+                                        style={styles.headerText}
+                                    >
+                                        Combinations in SYL Dark with contrast
+                                        issues
+                                    </BodyText>
+                                </View>
+                            </View>
+                            {passing.map((pair) => {
+                                const sdRatio = sdRatios.get(
+                                    `${pair.bgLabel}__${pair.fgLabel}`,
+                                );
+                                const sdRaw =
+                                    themes?.sd.backgrounds[pair.bgLabel];
+                                const sdTransparent = isTransparentValue(sdRaw);
+                                const sdFails =
+                                    sdTransparent ||
+                                    sdRatio == null ||
+                                    sdRatio < AA_THRESHOLD;
+                                const renderSyldarkTile = (
+                                    highlightFailing: boolean,
+                                ) =>
+                                    sdTransparent ? (
+                                        <View
+                                            style={[
+                                                styles.previewTile,
+                                                styles.previewTileEmpty,
+                                                highlightFailing &&
+                                                    styles.previewTileFailing,
+                                            ]}
+                                        >
+                                            <BodyText
+                                                size="small"
+                                                style={styles.emptyHint}
+                                            >
+                                                background.{pair.bgLabel} is
+                                                transparent in syl-dark; no
+                                                preview.
+                                            </BodyText>
+                                        </View>
+                                    ) : (
+                                        <ThemePreview
+                                            bgCssVar={pair.bgCssVar}
+                                            fgCssVar={pair.fgCssVar}
+                                            bgLabel={pair.bgLabel}
+                                            fgLabel={pair.fgLabel}
+                                            bgValue={
+                                                themes?.sd.backgrounds[
+                                                    pair.bgLabel
+                                                ]
+                                            }
+                                            fgValue={
+                                                themes?.sd.foregrounds[
+                                                    pair.fgLabel
+                                                ]
+                                            }
+                                            ratio={sdRatio ?? null}
+                                            highlightFailing={highlightFailing}
+                                        />
+                                    );
+                                return (
+                                    <View
+                                        key={`${pair.bgLabel}__${pair.fgLabel}`}
+                                        style={styles.darkRow}
+                                    >
+                                        {renderSyldarkTile(false)}
+                                        <View style={styles.failingSlot}>
+                                            {sdFails
+                                                ? renderSyldarkTile(true)
+                                                : null}
+                                        </View>
+                                    </View>
+                                );
+                            })}
+                        </View>
+                    </ThemeSwitcher>
+                </View>
+            </View>
+        );
+    },
+};
+
+// ---------------------------------------------------------------------------
+// Styles
+// ---------------------------------------------------------------------------
+
+const TILE_INLINE_SIZE = 420;
+const SWATCH_BLOCK_SIZE = 120;
 
 const swatchOutline = `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`;
 
 const styles = StyleSheet.create({
+    // --- Shared ---
     wrapper: {
         padding: sizing.size_240,
         backgroundColor: semanticColor.core.background.base.default,
@@ -559,6 +1016,8 @@ const styles = StyleSheet.create({
     loading: {
         padding: sizing.size_160,
     },
+
+    // --- ForegroundOnBackground grid ---
     grid: {
         display: "grid",
         columnGap: sizing.size_040,
@@ -608,8 +1067,6 @@ const styles = StyleSheet.create({
         textAlign: "center",
     },
     swatchTransparent: {
-        // Light checkerboard so the swatch reads as "transparent" without
-        // overpowering the variant label sitting on top.
         backgroundColor: semanticColor.core.background.base.default,
         backgroundImage: `conic-gradient(${semanticColor.core.background.neutral.subtle} 25%, transparent 0 50%, ${semanticColor.core.background.neutral.subtle} 0 75%, transparent 0)`,
         backgroundSize: `${sizing.size_080} ${sizing.size_080}`,
@@ -637,5 +1094,134 @@ const styles = StyleSheet.create({
     },
     cellRatioFail: {
         color: semanticColor.core.foreground.neutral.subtle,
+    },
+
+    // --- Cross-theme story ---
+    crossThemeWrapper: {
+        padding: sizing.size_240,
+        gap: sizing.size_240,
+        backgroundColor: semanticColor.core.background.base.default,
+    },
+    hiddenProbe: {
+        position: "absolute",
+        inlineSize: 0,
+        blockSize: 0,
+        overflow: "hidden",
+        visibility: "hidden",
+    },
+    summary: {
+        gap: sizing.size_080,
+        maxInlineSize: 720,
+    },
+    split: {
+        flexDirection: "row",
+        alignItems: "flex-start",
+        gap: sizing.size_160,
+    },
+    lightPanel: {
+        // Mirrors the dark panel's spacing so rows on each side align,
+        // but with no backdrop — the page background shows through.
+        gap: sizing.size_160,
+        padding: sizing.size_160,
+        borderRadius: border.radius.radius_080,
+    },
+    darkPanel: {
+        gap: sizing.size_160,
+        padding: sizing.size_160,
+        borderRadius: border.radius.radius_080,
+        // Resolves to syl-dark's page background since this panel is wrapped
+        // in <ThemeSwitcher theme="syl-dark">.
+        backgroundColor: semanticColor.core.background.base.default,
+    },
+    darkRow: {
+        flexDirection: "row",
+        alignItems: "stretch",
+        gap: sizing.size_160,
+    },
+    failingSlot: {
+        inlineSize: TILE_INLINE_SIZE,
+        flexShrink: 0,
+    },
+    headerCell: {
+        inlineSize: TILE_INLINE_SIZE,
+        flexShrink: 0,
+        paddingBlockEnd: sizing.size_080,
+    },
+    headerText: {
+        color: semanticColor.core.foreground.neutral.strong,
+    },
+    previewTile: {
+        boxSizing: "border-box",
+        inlineSize: TILE_INLINE_SIZE,
+        borderRadius: border.radius.radius_080,
+        border: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
+        overflow: "hidden",
+        backgroundColor: semanticColor.core.background.base.subtle,
+    },
+    previewTileEmpty: {
+        alignItems: "center",
+        justifyContent: "center",
+        padding: sizing.size_160,
+        minBlockSize: SWATCH_BLOCK_SIZE,
+    },
+    previewTileFailing: {
+        // Outline (not border) so the layout doesn't shift, drawing a bright
+        // ring around any tile that doesn't meet AA in its theme.
+        outline: `${border.width.thick} solid ${semanticColor.core.border.critical.strong}`,
+    },
+    emptyHint: {
+        color: semanticColor.core.foreground.neutral.subtle,
+        textAlign: "center",
+    },
+    previewSwatch: {
+        gap: sizing.size_080,
+        padding: sizing.size_160,
+        minBlockSize: SWATCH_BLOCK_SIZE,
+        justifyContent: "center",
+    },
+    previewSampleHeading: {
+        // Inherit text color from the swatch's `color` (the foreground token).
+        color: "inherit",
+    },
+    previewSampleBody: {
+        color: "inherit",
+    },
+    previewMeta: {
+        gap: sizing.size_040,
+        padding: sizing.size_120,
+        borderBlockStart: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
+        backgroundColor: semanticColor.core.background.base.default,
+    },
+    previewMetaRow: {
+        flexDirection: "row",
+        alignItems: "center",
+        gap: sizing.size_080,
+    },
+    tokenTable: {
+        display: "grid",
+        // The token-name column uses minmax(0, 1fr) so it shrinks (and
+        // ellipsizes) instead of overflowing past the meta's right padding
+        // when the path is long.
+        gridTemplateColumns: "auto minmax(0, 1fr) auto auto",
+        columnGap: sizing.size_120,
+        rowGap: sizing.size_040,
+        alignItems: "baseline",
+    },
+    tokenCellKind: {
+        color: semanticColor.core.foreground.neutral.subtle,
+        textTransform: "uppercase",
+        letterSpacing: "0.04em",
+    },
+    tokenCellName: {
+        color: semanticColor.core.foreground.neutral.strong,
+        minInlineSize: 0,
+        overflow: "hidden",
+        textOverflow: "ellipsis",
+        whiteSpace: "nowrap",
+    },
+    tokenCellMono: {
+        fontFamily: font.family.mono,
+        color: semanticColor.core.foreground.neutral.default,
+        whiteSpace: "nowrap",
     },
 });
