--- __docs__/wonder-blocks-tokens/cross-theme-contrast-pairs.stories.tsx
@@ -1,714 +1,0 @@
-import * as React from "react";
-import {StyleSheet} from "aphrodite";
-import {Meta, StoryObj} from "@storybook/react-vite";
-
-import checkCircleFillIcon from "@phosphor-icons/core/fill/check-circle-fill.svg";
-import xCircleFillIcon from "@phosphor-icons/core/fill/x-circle-fill.svg";
-
-import {View} from "@khanacademy/wonder-blocks-core";
-import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
-import {ThemeSwitcher} from "@khanacademy/wonder-blocks-theming";
-import {BodyText} from "@khanacademy/wonder-blocks-typography";
-import {
-    border,
-    font,
-    mix,
-    semanticColor,
-    sizing,
-} from "@khanacademy/wonder-blocks-tokens";
-
-import {color as thunderblocksPrimitives} from "../../packages/wonder-blocks-tokens/src/theme/semantic/internal/primitive-color-thunderblocks";
-import {flattenNestedTokens} from "../components/tokens-util";
-
-/**
- * Every `core.background` × `core.foreground` pair that passes WCAG **AA
- * normal text** (≥ 4.5 : 1) when resolved against the **thunderblocks**
- * theme. Each row shows the same token pair rendered in **thunderblocks** on
- * the left and **syl-dark** on the right so the two themes can be compared
- * side by side.
- *
- * Tokens with alpha are composited over `core.background.base.default` (per
- * theme) before the contrast is computed so the visible color drives the
- * decision.
- */
-export default {
-    title: "Internal / Cross-Theme Contrast Pairs",
-    parameters: {
-        chromatic: {disableSnapshot: true},
-    },
-    tags: ["!manifest"],
-} as Meta;
-
-type FlatToken = {label: string; cssVar: string};
-
-function flattenTokens(tokens: Record<string, unknown>): Array<FlatToken> {
-    const flat = flattenNestedTokens(tokens as any);
-    return Object.entries(flat).map(([label, cssVar]) => ({label, cssVar}));
-}
-
-const backgroundTokens = flattenTokens(semanticColor.core.background);
-const foregroundTokens = flattenTokens(semanticColor.core.foreground);
-
-type RGB = {r: number; g: number; b: number};
-
-function parseRgb(value: string): RGB | null {
-    if (!value) {
-        return null;
-    }
-    const trimmed = value.trim();
-    if (!trimmed || trimmed === "transparent") {
-        return null;
-    }
-
-    const hex6 = trimmed.match(/^#([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})$/i);
-    if (hex6) {
-        return {
-            r: parseInt(hex6[1], 16),
-            g: parseInt(hex6[2], 16),
-            b: parseInt(hex6[3], 16),
-        };
-    }
-    const hex3 = trimmed.match(/^#([0-9a-f])([0-9a-f])([0-9a-f])$/i);
-    if (hex3) {
-        return {
-            r: parseInt(hex3[1] + hex3[1], 16),
-            g: parseInt(hex3[2] + hex3[2], 16),
-            b: parseInt(hex3[3] + hex3[3], 16),
-        };
-    }
-    const rgb = trimmed.match(
-        /^rgba?\(\s*([\d.]+)\s*,\s*([\d.]+)\s*,\s*([\d.]+)/i,
-    );
-    if (rgb) {
-        return {
-            r: parseFloat(rgb[1]),
-            g: parseFloat(rgb[2]),
-            b: parseFloat(rgb[3]),
-        };
-    }
-    return null;
-}
-
-function relativeLuminance({r, g, b}: RGB): number {
-    const channel = (c: number) => {
-        const v = c / 255;
-        return v <= 0.03928 ? v / 12.92 : Math.pow((v + 0.055) / 1.055, 2.4);
-    };
-    return 0.2126 * channel(r) + 0.7152 * channel(g) + 0.0722 * channel(b);
-}
-
-function contrastRatio(a: string, b: string): number | null {
-    const aRgb = parseRgb(a);
-    const bRgb = parseRgb(b);
-    if (!aRgb || !bRgb) {
-        return null;
-    }
-    const la = relativeLuminance(aRgb);
-    const lb = relativeLuminance(bRgb);
-    const lighter = Math.max(la, lb);
-    const darker = Math.min(la, lb);
-    return (lighter + 0.05) / (darker + 0.05);
-}
-
-function compositeOver(value: string, base: string): string {
-    const v = (value || "").trim();
-    const b = (base || "").trim();
-    if (!v || v === "transparent") {
-        return b;
-    }
-    return mix(v, b || "#ffffff");
-}
-
-function isTransparentValue(value: string | undefined): boolean {
-    const v = (value || "").trim();
-    return !v || v === "transparent";
-}
-
-function formatRatio(ratio: number): string {
-    return ratio.toFixed(2).replace(/\.?0+$/, "");
-}
-
-function normalizeColorString(value: string): string {
-    return (value || "").trim().toLowerCase().replace(/\s+/g, "");
-}
-
-/**
- * Reverse map from a primitive's color value to its primitive name (e.g.
- * "#ffffff" → "white_100"). Both Thunderblocks and Syl-Dark draw from this
- * same primitive palette.
- */
-const PRIMITIVE_BY_VALUE = new Map<string, string>(
-    Object.entries(thunderblocksPrimitives).map(([name, value]) => [
-        normalizeColorString(value),
-        name,
-    ]),
-);
-
-function lookupPrimitiveName(value: string | undefined): string | null {
-    if (!value) {
-        return null;
-    }
-    return PRIMITIVE_BY_VALUE.get(normalizeColorString(value)) ?? null;
-}
-
-const AA_THRESHOLD = 4.5;
-
-type ResolvedTheme = {
-    backgrounds: Record<string, string>;
-    foregrounds: Record<string, string>;
-    pageBase: string;
-};
-
-function readThemeValues(el: HTMLElement | null): ResolvedTheme | null {
-    if (!el) {
-        return null;
-    }
-    const cs = getComputedStyle(el);
-    const resolve = (cssVar: string) => {
-        const match = cssVar.match(/^var\((--[^,)]+)/);
-        if (!match) {
-            return cssVar;
-        }
-        return cs.getPropertyValue(match[1]).trim() || cssVar;
-    };
-    return {
-        backgrounds: Object.fromEntries(
-            backgroundTokens.map(({label, cssVar}) => [label, resolve(cssVar)]),
-        ),
-        foregrounds: Object.fromEntries(
-            foregroundTokens.map(({label, cssVar}) => [label, resolve(cssVar)]),
-        ),
-        pageBase:
-            resolve(semanticColor.core.background.base.default) || "#ffffff",
-    };
-}
-
-type ResolvedPair = {
-    bgLabel: string;
-    fgLabel: string;
-    bgCssVar: string;
-    fgCssVar: string;
-    bgRaw: string;
-    fgRaw: string;
-    bgVisible: string;
-    fgVisible: string;
-    ratio: number;
-};
-
-function buildResolvedPairs(theme: ResolvedTheme): Array<ResolvedPair> {
-    const pairs: Array<ResolvedPair> = [];
-    for (const bg of backgroundTokens) {
-        const bgRaw = theme.backgrounds[bg.label];
-        if (isTransparentValue(bgRaw)) {
-            continue;
-        }
-        const bgVisible = compositeOver(bgRaw, theme.pageBase);
-        for (const fg of foregroundTokens) {
-            const fgRaw = theme.foregrounds[fg.label];
-            const fgVisible = compositeOver(fgRaw, bgVisible);
-            const ratio = contrastRatio(bgVisible, fgVisible);
-            if (ratio == null) {
-                continue;
-            }
-            pairs.push({
-                bgLabel: bg.label,
-                fgLabel: fg.label,
-                bgCssVar: bg.cssVar,
-                fgCssVar: fg.cssVar,
-                bgRaw,
-                fgRaw,
-                bgVisible,
-                fgVisible,
-                ratio,
-            });
-        }
-    }
-    return pairs;
-}
-
-function TokenRow({
-    kind,
-    tokenName,
-    value,
-}: {
-    kind: "bg" | "fg";
-    tokenName: string;
-    value: string | undefined;
-}) {
-    const trimmed = value?.trim() ?? "";
-    const primitive = lookupPrimitiveName(trimmed);
-    return (
-        <>
-            <BodyText
-                tag="span"
-                size="xsmall"
-                weight="bold"
-                style={styles.tokenCellKind}
-            >
-                {kind}
-            </BodyText>
-            <BodyText tag="span" size="xsmall" style={styles.tokenCellName}>
-                {tokenName}
-            </BodyText>
-            <BodyText tag="span" size="xsmall" style={styles.tokenCellMono}>
-                {primitive ?? "—"}
-            </BodyText>
-            <BodyText tag="span" size="xsmall" style={styles.tokenCellMono}>
-                {trimmed || "—"}
-            </BodyText>
-        </>
-    );
-}
-
-function ThemePreview({
-    bgCssVar,
-    fgCssVar,
-    bgLabel,
-    fgLabel,
-    bgValue,
-    fgValue,
-    ratio,
-    highlightFailing = false,
-}: {
-    bgCssVar: string;
-    fgCssVar: string;
-    bgLabel: string;
-    fgLabel: string;
-    /** The actual color string the bg CSS var resolves to in this theme. */
-    bgValue: string | undefined;
-    /** The actual color string the fg CSS var resolves to in this theme. */
-    fgValue: string | undefined;
-    ratio: number | null;
-    /** When true, draws the bright-red outline if the pair fails AA. */
-    highlightFailing?: boolean;
-}) {
-    const passes = ratio != null && ratio >= AA_THRESHOLD;
-    return (
-        <View
-            style={[
-                styles.previewTile,
-                highlightFailing && !passes && styles.previewTileFailing,
-            ]}
-        >
-            <View
-                style={[
-                    styles.previewSwatch,
-                    {backgroundColor: bgCssVar, color: fgCssVar},
-                ]}
-            >
-                <BodyText
-                    tag="span"
-                    size="medium"
-                    weight="bold"
-                    style={styles.previewSampleHeading}
-                >
-                    Aa Bb Cc
-                </BodyText>
-                <BodyText
-                    tag="span"
-                    size="small"
-                    style={styles.previewSampleBody}
-                >
-                    The quick brown fox jumps.
-                </BodyText>
-            </View>
-            <View style={styles.previewMeta}>
-                <View style={styles.previewMetaRow}>
-                    <PhosphorIcon
-                        icon={passes ? checkCircleFillIcon : xCircleFillIcon}
-                        size="small"
-                        color={
-                            passes
-                                ? semanticColor.core.foreground.success.default
-                                : semanticColor.core.foreground.neutral.subtle
-                        }
-                        aria-hidden={true}
-                    />
-                    <BodyText tag="span" size="xsmall">
-                        {ratio == null
-                            ? "no value"
-                            : `${formatRatio(ratio)} : 1`}{" "}
-                        · {passes ? "passes AA" : "fails AA"}
-                    </BodyText>
-                </View>
-                <View style={styles.tokenTable}>
-                    <TokenRow
-                        kind="bg"
-                        tokenName={`core.background.${bgLabel}`}
-                        value={bgValue}
-                    />
-                    <TokenRow
-                        kind="fg"
-                        tokenName={`core.foreground.${fgLabel}`}
-                        value={fgValue}
-                    />
-                </View>
-            </View>
-        </View>
-    );
-}
-
-export const PassingPairsThunderblocks: StoryObj = {
-    name: "Thunderblocks ↔ SylDark",
-    render: function Render() {
-        const tbRef = React.useRef<HTMLDivElement>(null);
-        const sdRef = React.useRef<HTMLDivElement>(null);
-        const [themes, setThemes] = React.useState<{
-            tb: ResolvedTheme;
-            sd: ResolvedTheme;
-        } | null>(null);
-
-        React.useEffect(() => {
-            const tb = readThemeValues(tbRef.current);
-            const sd = readThemeValues(sdRef.current);
-            if (tb && sd) {
-                setThemes({tb, sd});
-            }
-        }, []);
-
-        const passing = React.useMemo(() => {
-            if (!themes) {
-                return [];
-            }
-            return buildResolvedPairs(themes.tb).filter(
-                (pair) => pair.ratio >= AA_THRESHOLD,
-            );
-        }, [themes]);
-
-        const sdRatios = React.useMemo(() => {
-            if (!themes) {
-                return new Map<string, number | null>();
-            }
-            const map = new Map<string, number | null>();
-            for (const pair of passing) {
-                const bgRaw = themes.sd.backgrounds[pair.bgLabel];
-                const fgRaw = themes.sd.foregrounds[pair.fgLabel];
-                if (isTransparentValue(bgRaw)) {
-                    map.set(`${pair.bgLabel}__${pair.fgLabel}`, null);
-                    continue;
-                }
-                const bgVisible = compositeOver(bgRaw, themes.sd.pageBase);
-                const fgVisible = compositeOver(fgRaw, bgVisible);
-                map.set(
-                    `${pair.bgLabel}__${pair.fgLabel}`,
-                    contrastRatio(bgVisible, fgVisible),
-                );
-            }
-            return map;
-        }, [themes, passing]);
-
-        const sdFailingCount = React.useMemo(() => {
-            let n = 0;
-            for (const ratio of sdRatios.values()) {
-                if (ratio == null || ratio < AA_THRESHOLD) {
-                    n++;
-                }
-            }
-            return n;
-        }, [sdRatios]);
-
-        return (
-            <View style={styles.wrapper}>
-                {/* Hidden helpers used to read each theme's CSS variables. */}
-                <View style={styles.hiddenProbe}>
-                    <ThemeSwitcher theme="thunderblocks">
-                        <div ref={tbRef} />
-                    </ThemeSwitcher>
-                    <ThemeSwitcher theme="syl-dark">
-                        <div ref={sdRef} />
-                    </ThemeSwitcher>
-                </View>
-
-                <View style={styles.summary}>
-                    <BodyText size="medium" weight="bold">
-                        Passing AA pairs in thunderblocks
-                    </BodyText>
-                    {themes ? (
-                        <>
-                            <BodyText size="small">
-                                {passing.length} of{" "}
-                                {backgroundTokens.length *
-                                    foregroundTokens.length}{" "}
-                                background × foreground combinations meet ≥ 4.5
-                                : 1 in thunderblocks. Each row shows the same
-                                pair in syl-dark for comparison.
-                            </BodyText>
-                            <BodyText size="small">
-                                {sdFailingCount} of those {passing.length} pair
-                                {passing.length === 1 ? "" : "s"} fall below 4.5
-                                : 1 (or use a transparent background) in
-                                syl-dark.
-                            </BodyText>
-                        </>
-                    ) : (
-                        <BodyText size="small">
-                            Resolving theme tokens…
-                        </BodyText>
-                    )}
-                </View>
-
-                <View style={styles.split}>
-                    <ThemeSwitcher theme="thunderblocks">
-                        <View style={styles.lightColumn}>
-                            <View style={styles.headerCell}>
-                                <BodyText
-                                    tag="span"
-                                    size="medium"
-                                    weight="bold"
-                                    style={styles.headerText}
-                                >
-                                    Thunderblocks (SYL Light)
-                                </BodyText>
-                            </View>
-                            {passing.map((pair) => (
-                                <ThemePreview
-                                    key={`${pair.bgLabel}__${pair.fgLabel}`}
-                                    bgCssVar={pair.bgCssVar}
-                                    fgCssVar={pair.fgCssVar}
-                                    bgLabel={pair.bgLabel}
-                                    fgLabel={pair.fgLabel}
-                                    bgValue={pair.bgRaw}
-                                    fgValue={pair.fgRaw}
-                                    ratio={pair.ratio}
-                                />
-                            ))}
-                        </View>
-                    </ThemeSwitcher>
-                    <ThemeSwitcher theme="syl-dark">
-                        <View style={styles.darkColumn}>
-                            <View style={styles.darkRow}>
-                                <View style={styles.headerCell}>
-                                    <BodyText
-                                        tag="span"
-                                        size="medium"
-                                        weight="bold"
-                                        style={styles.headerText}
-                                    >
-                                        SYL Dark
-                                    </BodyText>
-                                </View>
-                                <View
-                                    style={[
-                                        styles.headerCell,
-                                        styles.failingSlot,
-                                    ]}
-                                >
-                                    <BodyText
-                                        tag="span"
-                                        size="medium"
-                                        weight="bold"
-                                        style={styles.headerText}
-                                    >
-                                        Combinations in SYL Dark with contrast
-                                        issues
-                                    </BodyText>
-                                </View>
-                            </View>
-                            {passing.map((pair) => {
-                                const sdRatio = sdRatios.get(
-                                    `${pair.bgLabel}__${pair.fgLabel}`,
-                                );
-                                const sdRaw =
-                                    themes?.sd.backgrounds[pair.bgLabel];
-                                const sdTransparent = isTransparentValue(sdRaw);
-                                const sdFails =
-                                    sdTransparent ||
-                                    sdRatio == null ||
-                                    sdRatio < AA_THRESHOLD;
-                                const renderSyldarkTile = (
-                                    highlightFailing: boolean,
-                                ) =>
-                                    sdTransparent ? (
-                                        <View
-                                            style={[
-                                                styles.previewTile,
-                                                styles.previewTileEmpty,
-                                                highlightFailing &&
-                                                    styles.previewTileFailing,
-                                            ]}
-                                        >
-                                            <BodyText
-                                                size="small"
-                                                style={styles.emptyHint}
-                                            >
-                                                background.{pair.bgLabel} is
-                                                transparent in syl-dark; no
-                                                preview.
-                                            </BodyText>
-                                        </View>
-                                    ) : (
-                                        <ThemePreview
-                                            bgCssVar={pair.bgCssVar}
-                                            fgCssVar={pair.fgCssVar}
-                                            bgLabel={pair.bgLabel}
-                                            fgLabel={pair.fgLabel}
-                                            bgValue={
-                                                themes?.sd.backgrounds[
-                                                    pair.bgLabel
-                                                ]
-                                            }
-                                            fgValue={
-                                                themes?.sd.foregrounds[
-                                                    pair.fgLabel
-                                                ]
-                                            }
-                                            ratio={sdRatio ?? null}
-                                            highlightFailing={highlightFailing}
-                                        />
-                                    );
-                                return (
-                                    <View
-                                        key={`${pair.bgLabel}__${pair.fgLabel}`}
-                                        style={styles.darkRow}
-                                    >
-                                        {renderSyldarkTile(false)}
-                                        <View style={styles.failingSlot}>
-                                            {sdFails
-                                                ? renderSyldarkTile(true)
-                                                : null}
-                                        </View>
-                                    </View>
-                                );
-                            })}
-                        </View>
-                    </ThemeSwitcher>
-                </View>
-            </View>
-        );
-    },
-};
-
-const TILE_INLINE_SIZE = 420;
-const SWATCH_BLOCK_SIZE = 120;
-
-const styles = StyleSheet.create({
-    wrapper: {
-        padding: sizing.size_240,
-        gap: sizing.size_240,
-        backgroundColor: semanticColor.core.background.base.default,
-    },
-    hiddenProbe: {
-        position: "absolute",
-        inlineSize: 0,
-        blockSize: 0,
-        overflow: "hidden",
-        visibility: "hidden",
-    },
-    summary: {
-        gap: sizing.size_080,
-        maxInlineSize: 720,
-    },
-    split: {
-        flexDirection: "row",
-        alignItems: "flex-start",
-        gap: sizing.size_160,
-    },
-    lightColumn: {
-        gap: sizing.size_160,
-        // Mirror the dark column's padding so the headers and rows in both
-        // columns line up horizontally.
-        padding: sizing.size_160,
-    },
-    darkColumn: {
-        gap: sizing.size_160,
-        padding: sizing.size_160,
-        borderRadius: border.radius.radius_080,
-        // Resolves to syl-dark's page background since this column is wrapped
-        // in <ThemeSwitcher theme="syl-dark">.
-        backgroundColor: semanticColor.core.background.base.default,
-    },
-    darkRow: {
-        flexDirection: "row",
-        alignItems: "stretch",
-        gap: sizing.size_160,
-    },
-    failingSlot: {
-        inlineSize: TILE_INLINE_SIZE,
-        flexShrink: 0,
-    },
-    headerCell: {
-        inlineSize: TILE_INLINE_SIZE,
-        flexShrink: 0,
-        paddingBlockEnd: sizing.size_080,
-    },
-    headerText: {
-        // Resolves to the surrounding theme's strongest neutral foreground —
-        // dark text on the light Thunderblocks header, light text on the
-        // syl-dark lane header.
-        color: semanticColor.core.foreground.neutral.strong,
-    },
-    previewTile: {
-        boxSizing: "border-box",
-        inlineSize: TILE_INLINE_SIZE,
-        borderRadius: border.radius.radius_080,
-        border: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
-        overflow: "hidden",
-        backgroundColor: semanticColor.core.background.base.subtle,
-    },
-    previewTileEmpty: {
-        alignItems: "center",
-        justifyContent: "center",
-        padding: sizing.size_160,
-        minBlockSize: SWATCH_BLOCK_SIZE,
-    },
-    previewTileFailing: {
-        // Outline (not border) so the layout doesn't shift, drawing a bright
-        // ring around any tile that doesn't meet AA in its theme.
-        outline: `${border.width.thick} solid ${semanticColor.core.border.critical.strong}`,
-    },
-    emptyHint: {
-        color: semanticColor.core.foreground.neutral.subtle,
-        textAlign: "center",
-    },
-    previewSwatch: {
-        gap: sizing.size_080,
-        padding: sizing.size_160,
-        minBlockSize: SWATCH_BLOCK_SIZE,
-        justifyContent: "center",
-    },
-    previewSampleHeading: {
-        // Inherit text color from the swatch's `color` (the foreground token).
-        color: "inherit",
-    },
-    previewSampleBody: {
-        color: "inherit",
-    },
-    previewMeta: {
-        gap: sizing.size_040,
-        padding: sizing.size_120,
-        borderBlockStart: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
-        backgroundColor: semanticColor.core.background.base.default,
-    },
-    previewMetaRow: {
-        flexDirection: "row",
-        alignItems: "center",
-        gap: sizing.size_080,
-    },
-    tokenTable: {
-        display: "grid",
-        // The token-name column uses minmax(0, 1fr) so it shrinks (and
-        // ellipsizes) instead of overflowing past the meta's right padding
-        // when the path is long.
-        gridTemplateColumns: "auto minmax(0, 1fr) auto auto",
-        columnGap: sizing.size_120,
-        rowGap: sizing.size_040,
-        alignItems: "baseline",
-    },
-    tokenCellKind: {
-        color: semanticColor.core.foreground.neutral.subtle,
-        textTransform: "uppercase",
-        letterSpacing: "0.04em",
-    },
-    tokenCellName: {
-        color: semanticColor.core.foreground.neutral.strong,
-        minInlineSize: 0,
-        overflow: "hidden",
-        textOverflow: "ellipsis",
-        whiteSpace: "nowrap",
-    },
-    tokenCellMono: {
-        fontFamily: font.family.mono,
-        color: semanticColor.core.foreground.neutral.default,
-        whiteSpace: "nowrap",
-    },
-});
